+++ parser/src/ast.rs
@@ -597,6 +597,7 @@ pub enum AlterTableDef {
     #[serde(rename = "A_Const")]
     Constant(Value),
     ReplicaIdentityStmt(Value),
+    SQLValueFunction(Value),
 }
 
 #[derive(Debug, Deserialize, Serialize)]
@@ -729,7 +730,7 @@ pub struct CreateStmt {
     /// relation to create
     pub relation: RelationKind,
     /// column definitions (list of ColumnDef)
-    #[serde(rename = "tableElts")]
+    #[serde(rename = "tableElts", default)]
     pub table_elts: Vec<TableElt>,
     /// relations to inherit from (list of inhRelation)
     #[serde(rename = "inhRelations")]