+++ parser/src/error.rs
@@ -1,12 +1,12 @@
 #[derive(Debug, PartialEq)]
-pub enum PGQueryError {
+pub enum PgQueryError {
     ParsingCString,
     JsonParse(String),
     QueryToCString,
-    PGParseError,
+    PgParseError,
 }
 
-impl std::fmt::Display for PGQueryError {
+impl std::fmt::Display for PgQueryError {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         match *self {
             Self::ParsingCString => write!(
@@ -19,24 +19,24 @@ impl std::fmt::Display for PGQueryError {
                 format!("Squawk schema failed to parse Postgres response. This indicates a bug with Squawk. Please report this error to https://github.com/sbdchd/squawk. Schema error: {}", err)
             ),
             Self::QueryToCString => write!(f, "Could not encode query into CString"),
-            Self::PGParseError => write!(f, "Postgres failed to parse query"),
+            Self::PgParseError => write!(f, "Postgres failed to parse query"),
         }
     }
 }
 
-impl std::convert::From<std::ffi::NulError> for PGQueryError {
+impl std::convert::From<std::ffi::NulError> for PgQueryError {
     fn from(_: std::ffi::NulError) -> Self {
         Self::QueryToCString
     }
 }
 
-impl std::convert::From<serde_json::error::Error> for PGQueryError {
+impl std::convert::From<serde_json::error::Error> for PgQueryError {
     fn from(e: serde_json::error::Error) -> Self {
         Self::JsonParse(e.to_string())
     }
 }
 
-impl std::convert::From<std::str::Utf8Error> for PGQueryError {
+impl std::convert::From<std::str::Utf8Error> for PgQueryError {
     fn from(_: std::str::Utf8Error) -> Self {
         Self::ParsingCString
     }