+++ CHANGELOG.md
@@ -7,14 +7,52 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## v0.8.0 - 2021-10-31
+
+### Fixed
+
+- false positives with disallowed-unique-constraint and adding-serial-primary-key-field. Thanks @qoelet! (#161)
+
+## v0.7.3 - 2021-08-11
+
+### Fixed
+
+- false positives with require-concurrent-index rule (#157)
+
+## v0.7.2 - 2021-08-02
+
+### Fixed
+
+- incorrect internal schema for "alter table" statements with function calls. Thanks @qoelet! (#154)
+
+## v0.7.1 - 2021-05-30
+
+### Fixed
+
+- incorrect internal schema for "create partition" statements. (#146)
+- `upload-to-github` command not obeying top level `--exclude`s. (#142)
+
+### Changed
+
+- allowing adding not null column with default for `adding-not-null-field`. (#144)
+
+## v0.7.0 - 2021-05-19
+
 ### Added
 
 - link to website for tty reporter when there are lint errors. (#120)
 - `DROP INDEX` support for "require-concurrent-index-creation" and "prefer-robust-stmts". (#124)
+- github comment now includes Squawk's version number. (#131)
+- new `ban-drop-column` rule (#132)
 
 ### Changed
 
 - updated "adding-not-null-field" to warn about making a column non-nullable with `NOT NULL`. See the ["adding-not-null-field" docs](https://squawkhq.com/docs/adding-not-nullable-field) for more information. (#101)
+
+### Fixed
+
+- false positive with `prefer-text-field` that wasn't allowing `varchar`
+  without a length specified
 
 ## v0.6.0 - 2021-02-19